Completed Activity Diagram for Use Case #1
User fills out a flashcard (could be valid or invalid), and is given the opportunity to either share the flashcard or take a quiz based on the information only if it is valid. Otherwise, they get another attempt to fix the invalid flashcard that they initially made. Additionally, if they take the quiz, they receive real-time feedback and can view quiz results at the end.